feat: 目标组批量接口的share_data里增加lb_id的数据
--story=116334330
