+++ cmd/cloud-server/service/load-balancer/tcloud_url_rule.go
@@ -393,6 +393,12 @@ func (svc *lbSvc) CreateBizTCloudUrlRule(cts *rest.Contexts) (any, error) {
 		return nil, err
 	}
 
+	// 预检测-是否有执行中的负载均衡
+	err = svc.checkResFlowRel(cts.Kit, lblInfo.LbID, enumor.LoadBalancerCloudResType)
+	if err != nil {
+		return nil, err
+	}
+
 	hcReq := &hcproto.TCloudRuleBatchCreateReq{Rules: []hcproto.TCloudRuleCreate{convRuleCreate(req, tg)}}
 	createResp, err := svc.client.HCService().TCloud.Clb.BatchCreateUrlRule(cts.Kit, lblID, hcReq)
 	if err != nil {
@@ -447,6 +453,7 @@ func (svc *lbSvc) applyTargetToRule(kt *kit.Kit, tgID, ruleCloudID string, lblIn
 		rsReq := &hcproto.BatchRegisterTCloudTargetReq{
 			CloudListenerID: lblInfo.CloudID,
 			CloudRuleID:     ruleCloudID,
+			TargetGroupID:   tgID,
 			RuleType:        ruleType,
 			Targets:         make([]*hcproto.RegisterTarget, 0, len(rsResp.Details)),
 		}
