+++ cmd/cloud-server/service/load-balancer/update.go
@@ -24,30 +24,22 @@ import (
 	"encoding/json"
 	"fmt"
 
-	actionlb "hcm/cmd/task-server/logics/action/load-balancer"
-	"hcm/cmd/task-server/logics/flow"
 	cloudserver "hcm/pkg/api/cloud-server"
 	cslb "hcm/pkg/api/cloud-server/load-balancer"
 	"hcm/pkg/api/core"
 	dataproto "hcm/pkg/api/data-service/cloud"
 	hclbproto "hcm/pkg/api/hc-service/load-balancer"
-	ts "hcm/pkg/api/task-server"
-	"hcm/pkg/async/action"
-	"hcm/pkg/criteria/constant"
 	"hcm/pkg/criteria/enumor"
 	"hcm/pkg/criteria/errf"
 	"hcm/pkg/dal/dao/tools"
 	"hcm/pkg/dal/dao/types"
-	tableasync "hcm/pkg/dal/table/async"
 	tabletype "hcm/pkg/dal/table/types"
 	"hcm/pkg/iam/meta"
 	"hcm/pkg/kit"
 	"hcm/pkg/logs"
 	"hcm/pkg/rest"
 	"hcm/pkg/tools/converter"
-	"hcm/pkg/tools/counter"
 	"hcm/pkg/tools/hooks/handler"
-	"hcm/pkg/tools/slice"
 )
 
 // UpdateBizTCloudLoadBalancer  业务下更新clb
@@ -109,7 +101,6 @@ func (svc *lbSvc) UpdateBizTargetGroup(cts *rest.Contexts) (any, error) {
 }
 
 func (svc *lbSvc) updateTargetGroup(cts *rest.Contexts, authHandler handler.ValidWithAuthHandler) (any, error) {
-
 	id := cts.PathParameter("id").String()
 	if len(id) == 0 {
 		return nil, errf.New(errf.InvalidParameter, "id is required")
@@ -140,16 +131,27 @@ func (svc *lbSvc) updateTargetGroup(cts *rest.Contexts, authHandler handler.Vali
 // 更新目标组基本信息
 func (svc *lbSvc) batchUpdateTCloudTargetGroup(cts *rest.Contexts, id string) (interface{}, error) {
 	req := new(cslb.TargetGroupUpdateReq)
-	if err := cts.DecodeInto(cts); err != nil {
+	if err := cts.DecodeInto(req); err != nil {
 		return nil, errf.NewFromErr(errf.DecodeRequestFailed, err)
 	}
 
 	if err := req.Validate(); err != nil {
 		return nil, errf.NewFromErr(errf.InvalidParameter, err)
 	}
+	req.IDs = append(req.IDs, id)
+
+	// 检查目标组是否已绑定RS，如已绑定则不能更新region、vpc
+	targetList, err := svc.getTargetByTGIDs(cts.Kit, req.IDs)
+	if err != nil {
+		return nil, err
+	}
+
+	if len(targetList) > 0 && (len(req.Region) > 0 || len(req.CloudVpcID) > 0) {
+		return nil, errf.New(errf.InvalidParameter, "target group has bind rs, region or vpc cannot be update")
+	}
 
 	dbReq := &dataproto.TargetGroupUpdateReq{
-		IDs:        append(req.IDs, id),
+		IDs:        req.IDs,
 		Name:       req.Name,
 		VpcID:      req.VpcID,
 		CloudVpcID: req.CloudVpcID,
@@ -158,7 +160,7 @@ func (svc *lbSvc) batchUpdateTCloudTargetGroup(cts *rest.Contexts, id string) (i
 		Port:       req.Port,
 		Weight:     req.Weight,
 	}
-	err := svc.client.DataService().TCloud.LoadBalancer.BatchUpdateTCloudTargetGroup(cts.Kit, dbReq)
+	err = svc.client.DataService().TCloud.LoadBalancer.BatchUpdateTCloudTargetGroup(cts.Kit, dbReq)
 	if err != nil {
 		logs.Errorf("update tcloud target group failed, req: %+v, err: %v, rid: %s", req, err, cts.Kit.Rid)
 		return nil, err
@@ -394,379 +396,4 @@ func (svc *lbSvc) updateDomainAttr(cts *rest.Contexts, authHandler handler.Valid
 	default:
 		return nil, fmt.Errorf("vendor: %s not support", baseInfo.Vendor)
 	}
-}
-
-// BatchModifyBizTargetsPort batch modify biz targets port.
-func (svc *lbSvc) BatchModifyBizTargetsPort(cts *rest.Contexts) (any, error) {
-	return svc.batchModifyTargetPort(cts, handler.BizOperateAuth)
-}
-
-func (svc *lbSvc) batchModifyTargetPort(cts *rest.Contexts,
-	authHandler handler.ValidWithAuthHandler) (any, error) {
-
-	tgID := cts.PathParameter("target_group_id").String()
-	if len(tgID) == 0 {
-		return nil, errf.New(errf.InvalidParameter, "target_group_id is required")
-	}
-
-	req := new(cloudserver.ResourceCreateReq)
-	if err := cts.DecodeInto(req); err != nil {
-		logs.Errorf("batch modify target port request decode failed, err: %v, rid: %s", err, cts.Kit.Rid)
-		return nil, errf.NewFromErr(errf.DecodeRequestFailed, err)
-	}
-
-	baseInfo, err := svc.client.DataService().Global.Cloud.GetResBasicInfo(
-		cts.Kit, enumor.TargetGroupCloudResType, tgID)
-	if err != nil {
-		logs.Errorf("get target group resource info failed, id: %s, err: %s, rid: %s", tgID, err, cts.Kit.Rid)
-		return nil, err
-	}
-
-	// authorized instances
-	err = authHandler(cts, &handler.ValidWithAuthOption{Authorizer: svc.authorizer, ResType: meta.TargetGroup,
-		Action: meta.Update, BasicInfo: baseInfo})
-	if err != nil {
-		logs.Errorf("batch modify target port auth failed, err: %v, rid: %s", err, cts.Kit.Rid)
-		return nil, err
-	}
-
-	switch baseInfo.Vendor {
-	case enumor.TCloud:
-		return svc.buildModifyTCloudTargetPort(cts.Kit, req.Data, tgID, baseInfo.AccountID)
-	default:
-		return nil, fmt.Errorf("vendor: %s not support", baseInfo.Vendor)
-	}
-}
-
-func (svc *lbSvc) buildModifyTCloudTargetPort(kt *kit.Kit, body json.RawMessage,
-	tgID, accountID string) (interface{}, error) {
-
-	req := new(cslb.TCloudBatchModifyTargetPortReq)
-	if err := json.Unmarshal(body, req); err != nil {
-		return nil, errf.NewFromErr(errf.DecodeRequestFailed, err)
-	}
-
-	if err := req.Validate(); err != nil {
-		return nil, errf.NewFromErr(errf.InvalidParameter, err)
-	}
-
-	// 根据目标组ID，获取目标组绑定的监听器、规则列表
-	ruleRelReq := &core.ListReq{
-		Filter: tools.EqualExpression("target_group_id", tgID),
-		Page:   core.NewDefaultBasePage(),
-	}
-	ruleRelList, err := svc.client.DataService().Global.LoadBalancer.ListTargetGroupListenerRel(kt, ruleRelReq)
-	if err != nil {
-		logs.Errorf("list tcloud listener url rule failed, tgID: %s, err: %v, rid: %s", tgID, err, kt.Rid)
-		return nil, err
-	}
-
-	// 该目标组尚未绑定监听器及规则，不需要云端操作
-	if len(ruleRelList.Details) == 0 {
-		if err = svc.batchUpdateTargetPortDb(kt, req); err != nil {
-			return nil, err
-		}
-		return &core.FlowStateResult{State: enumor.FlowSuccess}, nil
-	}
-
-	return svc.buildModifyTCloudTargetTasksPort(kt, req, ruleRelList.Details[0].LbID, tgID, accountID)
-}
-
-func (svc *lbSvc) batchUpdateTargetPortDb(kt *kit.Kit, req *cslb.TCloudBatchModifyTargetPortReq) error {
-	tgReq := &core.ListReq{
-		Filter: tools.ContainersExpression("id", req.TargetIDs),
-		Page:   core.NewDefaultBasePage(),
-	}
-	rsList, err := svc.client.DataService().Global.LoadBalancer.ListTarget(kt, tgReq)
-	if err != nil {
-		return err
-	}
-	if len(rsList.Details) == 0 {
-		return errf.Newf(errf.RecordNotFound, "target_ids: %v is not found", req.TargetIDs)
-	}
-
-	instExistsMap := make(map[string]struct{}, 0)
-	updateReq := &dataproto.TargetBatchUpdateReq{Targets: []*dataproto.TargetUpdate{}}
-	for _, item := range rsList.Details {
-		// 批量修改端口时，需要校验重复的实例ID的问题，否则云端接口也会报错
-		if _, ok := instExistsMap[item.CloudInstID]; ok {
-			return errf.Newf(errf.RecordDuplicated, "duplicate modify same inst(%s) to new_port", item.CloudInstID)
-		}
-
-		instExistsMap[item.CloudInstID] = struct{}{}
-		updateReq.Targets = append(updateReq.Targets, &dataproto.TargetUpdate{
-			ID:   item.ID,
-			Port: req.NewPort,
-		})
-	}
-
-	return svc.client.DataService().Global.LoadBalancer.BatchUpdateTarget(kt, updateReq)
-}
-
-func (svc *lbSvc) buildModifyTCloudTargetTasksPort(kt *kit.Kit, req *cslb.TCloudBatchModifyTargetPortReq, lbID, tgID,
-	accountID string) (interface{}, error) {
-
-	// 预检测
-	err := svc.checkResFlowRel(kt, lbID, enumor.LoadBalancerCloudResType)
-	if err != nil {
-		return nil, err
-	}
-
-	// 创建Flow跟Task的初始化数据
-	tasks := make([]ts.CustomFlowTask, 0)
-	elems := slice.Split(req.TargetIDs, constant.BatchModifyTargetPortCloudMaxLimit)
-	getActionID := counter.NewNumStringCounter(1, 10)
-	var lastActionID action.ActIDType
-	for _, parts := range elems {
-		rsPortParams, err := svc.convTCloudOperateTargetReq(kt, parts, tgID, accountID,
-			converter.ValToPtr(req.NewPort), nil)
-		if err != nil {
-			return nil, err
-		}
-		actionID := action.ActIDType(getActionID())
-		tmpTask := ts.CustomFlowTask{
-			ActionID:   actionID,
-			ActionName: enumor.ActionTargetGroupModifyPort,
-			Params: &actionlb.OperateRsOption{
-				Vendor:                      enumor.TCloud,
-				TCloudBatchOperateTargetReq: *rsPortParams,
-			},
-			Retry: &tableasync.Retry{
-				Enable: true,
-				Policy: &tableasync.RetryPolicy{
-					Count:        constant.FlowRetryMaxLimit,
-					SleepRangeMS: [2]uint{100, 200},
-				},
-			},
-		}
-		if len(lastActionID) > 0 {
-			tmpTask.DependOn = []action.ActIDType{lastActionID}
-		}
-		tasks = append(tasks, tmpTask)
-		lastActionID = actionID
-	}
-	removeReq := &ts.AddCustomFlowReq{Name: enumor.FlowTargetGroupModifyPort, Tasks: tasks, IsInitState: true}
-	result, err := svc.client.TaskServer().CreateCustomFlow(kt, removeReq)
-	if err != nil {
-		logs.Errorf("call taskserver to batch modify target port custom flow failed, err: %v, rid: %s", err, kt.Rid)
-		return nil, err
-	}
-
-	flowID := result.ID
-	// 从Flow，负责监听主Flow的状态
-	flowWatchReq := &ts.AddTemplateFlowReq{
-		Name: enumor.FlowLoadBalancerOperateWatch,
-		Tasks: []ts.TemplateFlowTask{{
-			ActionID: "1",
-			Params: &actionflow.LoadBalancerOperateWatchOption{
-				FlowID:     flowID,
-				ResID:      lbID,
-				ResType:    enumor.LoadBalancerCloudResType,
-				SubResIDs:  []string{tgID},
-				SubResType: enumor.TargetGroupCloudResType,
-				TaskType:   enumor.ModifyPortTaskType,
-			},
-		}},
-	}
-	_, err = svc.client.TaskServer().CreateTemplateFlow(kt, flowWatchReq)
-	if err != nil {
-		logs.Errorf("call taskserver to create res flow status watch task failed, err: %v, flowID: %s, rid: %s",
-			err, flowID, kt.Rid)
-		return nil, err
-	}
-
-	// 锁定资源跟Flow的状态
-	err = svc.lockResFlowStatus(kt, lbID, enumor.LoadBalancerCloudResType, flowID, enumor.ModifyPortTaskType)
-	if err != nil {
-		return nil, err
-	}
-
-	return &core.FlowStateResult{FlowID: flowID}, nil
-}
-
-// BatchModifyBizTargetsWeight batch modify biz targets weight.
-func (svc *lbSvc) BatchModifyBizTargetsWeight(cts *rest.Contexts) (any, error) {
-	return svc.batchModifyTargetWeight(cts, handler.BizOperateAuth)
-}
-
-func (svc *lbSvc) batchModifyTargetWeight(cts *rest.Contexts, authHandler handler.ValidWithAuthHandler) (any, error) {
-	tgID := cts.PathParameter("target_group_id").String()
-	if len(tgID) == 0 {
-		return nil, errf.New(errf.InvalidParameter, "target_group_id is required")
-	}
-
-	req := new(cloudserver.ResourceCreateReq)
-	if err := cts.DecodeInto(req); err != nil {
-		logs.Errorf("batch modify target weight request decode failed, err: %v, rid: %s", err, cts.Kit.Rid)
-		return nil, errf.NewFromErr(errf.DecodeRequestFailed, err)
-	}
-
-	baseInfo, err := svc.client.DataService().Global.Cloud.GetResBasicInfo(
-		cts.Kit, enumor.TargetGroupCloudResType, tgID)
-	if err != nil {
-		logs.Errorf("get target group resource info failed, id: %s, err: %s, rid: %s", tgID, err, cts.Kit.Rid)
-		return nil, err
-	}
-
-	// authorized instances
-	err = authHandler(cts, &handler.ValidWithAuthOption{Authorizer: svc.authorizer, ResType: meta.TargetGroup,
-		Action: meta.Update, BasicInfo: baseInfo})
-	if err != nil {
-		logs.Errorf("batch modify target weight auth failed, err: %v, rid: %s", err, cts.Kit.Rid)
-		return nil, err
-	}
-
-	switch baseInfo.Vendor {
-	case enumor.TCloud:
-		return svc.buildModifyTCloudTargetWeight(cts.Kit, req.Data, tgID, baseInfo.AccountID)
-	default:
-		return nil, fmt.Errorf("vendor: %s not support", baseInfo.Vendor)
-	}
-}
-
-func (svc *lbSvc) buildModifyTCloudTargetWeight(kt *kit.Kit, body json.RawMessage,
-	tgID, accountID string) (interface{}, error) {
-
-	req := new(cslb.TCloudBatchModifyTargetWeightReq)
-	if err := json.Unmarshal(body, req); err != nil {
-		return nil, errf.NewFromErr(errf.DecodeRequestFailed, err)
-	}
-
-	if err := req.Validate(); err != nil {
-		return nil, errf.NewFromErr(errf.InvalidParameter, err)
-	}
-
-	// 根据目标组ID，获取目标组绑定的监听器、规则列表
-	ruleRelReq := &core.ListReq{
-		Filter: tools.EqualExpression("target_group_id", tgID),
-		Page:   core.NewDefaultBasePage(),
-	}
-	ruleRelList, err := svc.client.DataService().Global.LoadBalancer.ListTargetGroupListenerRel(kt, ruleRelReq)
-	if err != nil {
-		logs.Errorf("list tcloud listener url rule failed, tgID: %s, err: %v, rid: %s", tgID, err, kt.Rid)
-		return nil, err
-	}
-
-	// 该目标组尚未绑定监听器及规则，不需要云端操作
-	if len(ruleRelList.Details) == 0 {
-		err = svc.batchUpdateTargetWeightDb(kt, req)
-		if err != nil {
-			return nil, err
-		}
-		return &core.FlowStateResult{State: enumor.FlowSuccess}, nil
-	}
-
-	return svc.buildModifyTCloudTargetTasksWeight(kt, req, ruleRelList.Details[0].LbID, tgID, accountID)
-}
-
-func (svc *lbSvc) batchUpdateTargetWeightDb(kt *kit.Kit, req *cslb.TCloudBatchModifyTargetWeightReq) error {
-	tgReq := &core.ListReq{
-		Filter: tools.ContainersExpression("id", req.TargetIDs),
-		Page:   core.NewDefaultBasePage(),
-	}
-	rsList, err := svc.client.DataService().Global.LoadBalancer.ListTarget(kt, tgReq)
-	if err != nil {
-		return err
-	}
-	if len(rsList.Details) == 0 {
-		return errf.Newf(errf.RecordNotFound, "target_ids: %v is not found", req.TargetIDs)
-	}
-
-	instExistsMap := make(map[string]struct{}, 0)
-	updateReq := &dataproto.TargetBatchUpdateReq{Targets: []*dataproto.TargetUpdate{}}
-	for _, item := range rsList.Details {
-		// 批量修改端口时，需要校验重复的实例ID的问题，否则云端接口也会报错
-		if _, ok := instExistsMap[item.CloudInstID]; ok {
-			return errf.Newf(errf.RecordDuplicated, "duplicate modify same inst(%s) to new_port", item.CloudInstID)
-		}
-
-		instExistsMap[item.CloudInstID] = struct{}{}
-		updateReq.Targets = append(updateReq.Targets, &dataproto.TargetUpdate{
-			ID:     item.ID,
-			Weight: converter.ValToPtr(req.NewWeight),
-		})
-	}
-
-	return svc.client.DataService().Global.LoadBalancer.BatchUpdateTarget(kt, updateReq)
-}
-
-func (svc *lbSvc) buildModifyTCloudTargetTasksWeight(kt *kit.Kit, req *cslb.TCloudBatchModifyTargetWeightReq,
-	lbID, tgID, accountID string) (interface{}, error) {
-
-	// 预检测
-	err := svc.checkResFlowRel(kt, lbID, enumor.LoadBalancerCloudResType)
-	if err != nil {
-		return nil, err
-	}
-
-	// 创建Flow跟Task的初始化数据
-	tasks := make([]ts.CustomFlowTask, 0)
-	elems := slice.Split(req.TargetIDs, constant.BatchModifyTargetWeightCloudMaxLimit)
-	getActionID := counter.NewNumStringCounter(1, 10)
-	var lastActionID action.ActIDType
-	for _, parts := range elems {
-		rsWeightParams, err := svc.convTCloudOperateTargetReq(kt, parts, tgID, accountID,
-			nil, converter.ValToPtr(req.NewWeight))
-		if err != nil {
-			return nil, err
-		}
-		actionID := action.ActIDType(getActionID())
-		tmpTask := ts.CustomFlowTask{
-			ActionID:   actionID,
-			ActionName: enumor.ActionTargetGroupModifyWeight,
-			Params: &actionlb.OperateRsOption{
-				Vendor:                      enumor.TCloud,
-				TCloudBatchOperateTargetReq: *rsWeightParams,
-			},
-			Retry: &tableasync.Retry{
-				Enable: true,
-				Policy: &tableasync.RetryPolicy{
-					Count:        constant.FlowRetryMaxLimit,
-					SleepRangeMS: [2]uint{100, 200},
-				},
-			},
-		}
-		if len(lastActionID) > 0 {
-			tmpTask.DependOn = []action.ActIDType{lastActionID}
-		}
-		tasks = append(tasks, tmpTask)
-		lastActionID = actionID
-	}
-	rsWeightReq := &ts.AddCustomFlowReq{Name: enumor.FlowTargetGroupModifyWeight, Tasks: tasks, IsInitState: true}
-	result, err := svc.client.TaskServer().CreateCustomFlow(kt, rsWeightReq)
-	if err != nil {
-		logs.Errorf("call taskserver to batch modify target weight custom flow failed, err: %v, rid: %s", err, kt.Rid)
-		return nil, err
-	}
-
-	flowID := result.ID
-	// 从Flow，负责监听主Flow的状态
-	flowWatchReq := &ts.AddTemplateFlowReq{
-		Name: enumor.FlowLoadBalancerOperateWatch,
-		Tasks: []ts.TemplateFlowTask{{
-			ActionID: "1",
-			Params: &actionflow.LoadBalancerOperateWatchOption{
-				FlowID:     flowID,
-				ResID:      lbID,
-				ResType:    enumor.LoadBalancerCloudResType,
-				SubResIDs:  []string{tgID},
-				SubResType: enumor.TargetGroupCloudResType,
-				TaskType:   enumor.ModifyWeightTaskType,
-			},
-		}},
-	}
-	_, err = svc.client.TaskServer().CreateTemplateFlow(kt, flowWatchReq)
-	if err != nil {
-		logs.Errorf("call taskserver to create res flow status watch task failed, err: %v, flowID: %s, rid: %s",
-			err, flowID, kt.Rid)
-		return nil, err
-	}
-
-	// 锁定资源跟Flow的状态
-	err = svc.lockResFlowStatus(kt, lbID, enumor.LoadBalancerCloudResType, flowID, enumor.ModifyWeightTaskType)
-	if err != nil {
-		return nil, err
-	}
-
-	return &core.FlowStateResult{FlowID: flowID}, nil
 }
