+++ docs/api-docs/web-server/docs/biz/load-balancer/get_record_async_task.md
@@ -41,6 +41,9 @@ POST /api/v1/cloud/bizs/{bk_biz_id}/audits/async_task/list
         "reason": {
             "message": "some tasks failed to be executed"
         },
+        "share_data": {
+            "lb_id": "00000001"
+        },
         "creator": "Jim",
         "reviser": "Jim",
         "created_at": "2023-02-12T14:47:39Z",
@@ -107,12 +110,17 @@ POST /api/v1/cloud/bizs/{bk_biz_id}/audits/async_task/list
 | created_at     | string  | 子任务创建时间，标准格式：2006-01-02T15:04:05Z            |
 | updated_at     | string  | 子任务最后一次修改时间，标准格式：2006-01-02T15:04:05Z     |
 
-
 #### data.flow.reason
 
 | 参数名称   | 参数类型 | 描述    |
 |----------|---------|---------|
 | message  | string  | 任务失败原因 |
+
+#### data.flow.share_data
+
+| 参数名称   | 参数类型 | 描述      |
+|----------|---------|-----------|
+| lb_id    | string  | 负载均衡ID |
 
 #### data.tasks[n].reason
